MP_SILSIM: more fixes
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_SIL@2041 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ Tools/MatrixPilot-SIL/MatrixPilot-SIL/SIL-udb.h
@@ -12,13 +12,13 @@
 #include "SIL-dsp.h"
 
 // Fake ASM calls
-#define __builtin_mulss(x,y) ((long)((short)(x)*(short)(y)))
-#define __builtin_mulus(x,y) ((long)((unsigned short)(x)*(short)(y)))
-#define __builtin_mulsu(x,y) ((long)((short)(x)*(unsigned short)(y)))
-#define __builtin_muluu(x,y) ((unsigned long)((unsigned short)(x)*(unsigned short)(y)))
+#define __builtin_mulss(x,y) ((( int32_t)(x))*( int32_t)(y))
+#define __builtin_mulus(x,y) (((uint32_t)(x))*( int32_t)(y))
+#define __builtin_mulsu(x,y) ((( int32_t)(x))*(uint32_t)(y))
+#define __builtin_muluu(x,y) (((uint32_t)(x))*(uint32_t)(y))
 
-#define __builtin_divud(x,y) ((short)((unsigned long)(x)/(short)(y)))
-#define __builtin_divsd(x,y) ((short)((long)(x)/(short)(y)))
+#define __builtin_divud(x,y) ((int16_t)(((uint32_t)(x))/(int16_t)(y)))
+#define __builtin_divsd(x,y) ((int16_t)((( int32_t)(x))/(int16_t)(y)))
 
 
 // LEDs
